backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-modeltranslation>=0.18
django-filter==23.5
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
Pillow>=10.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
pydantic>=2.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.34.0
httpx>=0.25.0
requests>=2.31.0
PyMuPDF>=1.23.0
# Environment variables
python-dotenv>=1.0.0
# Celery + Redis
celery>=5.3.0
redis>=5.0.0
# S3 / MinIO
boto3>=1.34.0
django-storages[boto3]>=1.14.0
# File type detection (magic bytes validation)
python-magic>=0.4.27
# Image hashing (дедупликация медиа)
imagehash>=4.3.0
# ElevenLabs (транскрибация голосовых — Scribe v2)
elevenlabs>=1.0.0
# Banking encryption
cryptography>=42.0.0
# Webhook JWT verification
PyJWT>=2.8.0
# Static files (production)
whitenoise>=6.0.0
# Monitoring
sentry-sdk[django,celery]>=1.40.0
# Testing
pytest>=8.0.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
factory-boy>=3.3.0